pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "issue-tracker"
version = "0.1.0"
description = "Issue tracker with external API integrations"
dependencies = [
    "django>=4.2,<5.0",
    "markdown>=3.4.0",
    "slack-sdk",
    "PyGithub",
    "python-dotenv",
    "whitenoise>=6.0.0",
]
requires-python = ">=3.8"

[project.scripts]
validate-configuration = "src.external_apis.validate_configuration:validate_configuration"

[tool.hatch.build.targets.wheel]
packages = ["src"]